--- modi/seller-service/src/main/java/com/coc/modi/seller/application/port/RentalPort.java
@@ -1,14 +1,0 @@
-package com.coc.modi.seller.application.port;
-
-import com.coc.modi.seller.infrastructure.client.rental.dto.RentalListResponse;
-
-public interface RentalPort {
-
-    RentalListResponse getRentals(Long sellerId,
-                                  String status,
-                                  String startDate,
-                                  String endDate,
-                                  Long productId,
-                                  Integer page,
-                                  Integer size);
-}
