+++ modi/seller-service/src/main/java/com/coc/modi/seller/seller/application/SellerService.java
@@ -1,7 +1,6 @@
 package com.coc.modi.seller.seller.application;
 
-import com.coc.modi.seller.application.SellerRentalService;
-import com.coc.modi.seller.application.dto.SellerRentalResponse;
+import com.coc.modi.seller.seller.application.dto.SellerRentalResponse;
 import com.coc.modi.seller.exception.SellerDuplicateException;
 import com.coc.modi.seller.exception.SellerNotFoundException;
 import com.coc.modi.seller.seller.application.dto.SellerCreateCommand;
