+++ modi/seller-service/src/main/java/com/coc/modi/seller/seller/infrastructure/client/rental/RentalClient.java
@@ -1,14 +1,19 @@
-package com.coc.modi.seller.infrastructure.client.rental;
+package com.coc.modi.seller.seller.infrastructure.client.rental;
 
-import com.coc.modi.seller.infrastructure.client.rental.dto.RentalListResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
-@FeignClient(name = "rentalClient", url = "${rental-service.url}")
+import com.coc.modi.seller.seller.infrastructure.client.rental.dto.RentalListResponse;
+
+@FeignClient(
+        name = "rental-service",
+        url = "${rental-service.url}",
+        path = "/internal/rentals"
+)
 public interface RentalClient {
 
-    @GetMapping("/internal/rentals")
+    @GetMapping
     RentalListResponse getRentals(@RequestParam("sellerId") Long sellerId,
                                   @RequestParam(value = "productId", required = false) Long productId,
                                   @RequestParam("status") String status,
