+++ modi/seller-service/src/main/java/com/coc/modi/seller/seller/infrastructure/client/rental/dto/RentalItemInfo.java
@@ -1,4 +1,4 @@
-package com.coc.modi.seller.infrastructure.client.rental.dto;
+package com.coc.modi.seller.seller.infrastructure.client.rental.dto;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
