+++ modi/seller-service/src/main/java/com/coc/modi/seller/seller/infrastructure/client/rental/dto/RentalListResponse.java
@@ -1,4 +1,4 @@
-package com.coc.modi.seller.infrastructure.client.rental.dto;
+package com.coc.modi.seller.seller.infrastructure.client.rental.dto;
 
 import java.util.List;
