+++ modi/seller-service/src/main/java/com/coc/modi/seller/settlement/batch/SettlementAggregationProcessor.java
@@ -1,6 +1,6 @@
 package com.coc.modi.seller.settlement.batch;
 
-import com.coc.modi.seller.infrastructure.client.rental.dto.RentalItemInfo;
+import com.coc.modi.seller.seller.infrastructure.client.rental.dto.RentalItemInfo;
 import com.coc.modi.seller.exception.SettlementPeriodResolveException;
 
 import org.springframework.batch.item.ItemProcessor;
